+++ jest.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  collectCoverage: true,
-  coverageDirectory: 'coverage',
-  coverageReporters: ['lcov'],
-  testURL: 'http://localhost/',
-  watchman: false,
-}